btrfs-progs: merge find-root.c and btrfs-find-root.c
